+++ core/src/main/java/com/genericgames/samurai/combat/ChargeAttack.java
@@ -4,9 +4,9 @@ import com.genericgames.samurai.model.movable.State;
 
 public class ChargeAttack extends Attack {
 
-    private float chargeDuration;
+    private int chargeDuration;
 
-    public ChargeAttack(float duration, float chargeDuration, int strength, State state) {
+    public ChargeAttack(int duration, int chargeDuration, int strength, State state) {
         super(duration, strength, state);
         this.chargeDuration = chargeDuration;
     }
@@ -15,7 +15,7 @@ public class ChargeAttack extends Attack {
         return chargeDuration;
     }
 
-    public void setChargeDuration(float chargeDuration) {
+    public void setChargeDuration(int chargeDuration) {
         this.chargeDuration = chargeDuration;
     }
 }
